requirements.txt
playwright
pandas
openpyxl
python-dotenv
rich
pyinstaller
tkcalendar
